Fixed temporary functionality for ticket button selector, this will be used for future work with johannes in cinemaSeating.
